+++ src/routes/trainingCycle/[id]/+page.server.ts
@@ -1,9 +1,9 @@
 import { getSignedUrlPromises } from '$lib/aws/s3';
 import { APIError } from '$lib/errors';
 import { prisma } from '$lib/prisma';
-import { TrainingCycleRepo, trainingCycleSchema } from '$lib/trainingCycle';
+import { TrainingCycleRepo } from '$lib/trainingCycle';
 import { getSessionOrRedirect } from '$lib/utils';
-import { exerciseGroupSchema } from '$lib/zodSchemas';
+import { exerciseGroupSchema, trainingCycleSchema } from '$lib/zodSchemas';
 import { fail } from '@sveltejs/kit';
 import type { Crumb } from 'svelte-breadcrumbs';
 import { superValidate } from 'sveltekit-superforms/server';
